+++ src/bible/index.ts
@@ -1,0 +1,7 @@
+import BibleReference from './BibleReference'
+import extractReferences from './extractReferences'
+
+export default {
+  BibleReference,
+  extractReferences
+}