+++ src/index.ts
@@ -2,9 +2,10 @@ import discord from 'discord.js'
 import 'dotenv/config'
 import commands from './commands'
 import config from './config'
+import bible from './bible'
 
 const client = new discord.Client({
-  intents: ['Guilds']
+  intents: ['Guilds', 'GuildMessages', 'MessageContent']
 })
 
 client.on('guildCreate', async guild => {
@@ -13,6 +14,22 @@ client.on('guildCreate', async guild => {
   if (typeof member === 'undefined') return
 
   member.setNickname(config.nickname)
+})
+
+client.on('messageCreate', async message => {
+  if (message.author.bot) return
+
+  const references = bible.extractReferences(message.content)
+
+  if (references.length === 0) return
+
+  const promises = references.map(reference => {
+    return message.channel.send(
+      reference.quote({ form: 'embed', inline: false })
+    )
+  })
+
+  await Promise.all(promises)
 })
 
 client.on('ready', async readyClient => {
@@ -25,4 +42,4 @@ client.on('interactionCreate', async interaction => {
   commands.handle(interaction)
 })
 
-client.login()
+client.login(process.env['DISCORD_TOKEN'])