+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `/changelog` command to show this file.
 - set nickname to `PBDB` when joining a server.
+- post Bible verses referenced in messages.
 
 ---